The diff contains changes to the .gitignore and src/pages/index.tsx files.
In the .gitignore file, a new line has been added for .env.

In the src/pages/index.tsx file, a few changes have been made. The import for Inter has been commented out, as well as the Inter constant. In addition, the contents of the <h2> and <p> tags within the <a> tags have been commented out.