This diff is for the file src/pages/api/serp/scholar.ts.
The first change is on line 6, where a new line is being added. This new line adds an object key to the url search params.

The second change is on line 10, where a new line is being added. This new line adds the value of the req.body to the url search params.