The first file, SearchBox.tsx, has a few changes. First, a line is removed that imports axios. Then, a few lines are added. These lines add a search icon from the React-icons library, and they also add a function that will be called when the search icon is clicked. This function will redirect the user to the search page with the query that they entered.
The second file, index.tsx, has a few changes as well. First, a line is added that imports the GetServerSideProps function from the next library. This function will be used to fetch data from the server before the page is rendered. Next, a few lines are added to the existing getServerSideProps function. These lines fetch the query from the context and return it as a prop to the page. Finally, a new line is added at the end of the file, which exports the getServerSideProps function.