The diff shows the changes made to the 'src/pages/scholar/index.tsx' file.
The first change is on line 29, where the 'Router' import is replaced with an 'axios' import.

The second change is on line 50, where a new prop is added to the 'ScholarPage' component. This prop is called 'searchResult', and it is used to store the results of a search made using the 'axios' module.

The third and final change is on line 64, where the 'structuredClone' function is used to clone the 'serpRes.data.organic_results' object. This ensures that the original data is not modified when it is passed to the 'searchResult' prop.