The first thing to note is that this diff is for multiple files. The main file being changed is index.tsx, but there are also changes to SearchBox.tsx and ResultSkeleton.tsx.
In the index.tsx file, the biggest change is the addition of the ScholarResultSkeleton component. This component is used to display a skeleton of the search results, which is useful for when the results are loading.

The SearchBox component has also been changed, but it's not clear what the changes are. It's possible that the changes are related to the ResultSkeleton component, but it's not certain.