The first file in this diff is the ResultSkeleton component, which is responsible for displaying a list of results from a search. The second file is the ScholarPage component, which is responsible for displaying the ScholarResultSkeleton component.
In the ResultSkeleton component, the organicResults prop is being destructured from the props object. The organicResults variable is then being mapped over and the ResultCard component is being rendered for each result.

In the ScholarPage component, the searchInfo and organicResults props are being destructured from the props object. The searchInfo variable contains information about the search, such as the total number of results and the time taken to perform the search. The organicResults variable contains the actual search results.