The first change is to the className attribute of the SearchBox component. The className is changed from flex flex-row w-full lg:mt-8 lg:w-2/3 mx-auto to flex flex-row w-full lg:w-2/3. This will change the way the SearchBox component is displayed on the page.
The second change is to the layout component. The classNames function is imported and used to add a className to the nav element. The className added is justify-between, which will change the way the nav element is displayed on the page.

The third change is to the ScholarPage component. The className of the container element is changed from container mx-auto p-4 to container mx-auto flex flex-col items-center justify-center p-4. This will change the way the ScholarPage component is displayed on the page.