The first thing to notice is that the path for the Footer component has changed. In the old version, the Footer was imported from the @/common/Footer path, but in the new version, it is imported from @/common/components/Footer. This suggests that either the Footer has been moved to a new location, or the path has been updated to point to a new location.
The second thing to notice is that the ScholarSearchBox component has been added. This component doesn't seem to have any changes, so it's likely that it's just been added to the project.

Finally, there are some changes to the Layout component itself. The biggest change is that the FaBuffer and FaUserCircle icons have been removed. These were probably being used for some specific purpose that is no longer needed. There are also some small changes to the way the component is rendered, but nothing major.