The first file in the diff, ResultPaginationBtn.tsx, has had some changes made to the way that pagination is displayed. Previously, the pagination buttons were just plain text links. Now, they are styled with CSS to look like buttons, and the current page is highlighted.
The second file, ResultSkeleton.tsx, has had a change made to the props that are passed to the ScholarResultPaginationBtn component. Previously, only the searchInfo and organicResults props were passed. Now, the pagination prop is also being passed.

The third and final file in the diff, scholar/index.tsx, has had a small change made to the ScholarResultSkeleton component. Previously, this component was only being passed the searchInfo and organicResults props. Now, it is also being passed the pagination prop.