The first file that is being changed is the Footer.tsx. In this file, the className for the footer is being changed from p-4 flex flex-row items-center justify-center to p-4 flex flex-row items-center justify-center border-t. This is a simple change that is just adding a border to the top of the footer element.
The second file that is being changed is the ResultCard.tsx. In this file, the code is being changed to make sure that the link element always has a href attribute. This is being done by adding a ?? after the result.link variable. This will make sure that if there is no link, then the href will just be an empty string.

The third file that is being changed is the ResultPaginationBtn.tsx. In this file, the code is being changed to create a link for each page number. This is being done by using the assignPageQueryToURL function. This function will take the base URL and the query and return a URL with the query string added to it. The query string will be added to the URL with the page number that is being passed in.

The fourth file that is being changed is the ResultSkeleton.tsx. In this file, the code is being changed to make sure that the query is always passed down to the ResultPaginationBtn component. This is being done by adding the query prop to the ResultPaginationBtn component.