The first diff is for the file ResultHeader.tsx. The main change here is that a new prop called query has been added to the ScholarResultHeader component. This prop will be used to pass the query string to the ResultHeaderFilter component. Additionally, a new state variable called showFilter has been added. This variable is used to control whether or not the ResultHeaderFilter component is displayed.
The second diff is for the file ResultHeaderFilter.tsx. The main change here is that the createFilterLink function now takes a string as the second argument instead of an number. This is necessary to properly create the filter links.

The third diff is for the file ResultSkeleton.tsx. The main change here is that the query prop is now being passed to the ScholarResultHeader component. This is necessary to properly display the ResultHeaderFilter component.