The first change in this diff is to the title of the ResultCard component. The title is now a link to the result that is being displayed.
The second change is to the way that the Cited by total is displayed. In the new code, the total is wrapped in a link to the citation page.

The third change is to the way that the author information is displayed. In the new code, the author's name is a link to their author page.