The first change is to the ResultCard component. The className prop has been changed from p-4 lg:px-0 border-b border-white lg:w-2/3 to p-4 lg:pl-0 border-b border-white. This will change the padding on the left side of the component for large screens.
The second change is to the ResultFilter component. The className prop has been changed from p-4 border-b border-blue-200 to p-4 border-b border-blue-200 hover:bg-black hover:text-white. This will change the background color and text color when the user hovers over the component.

The third change is to the ResultSkeleton component. The component has been changed to use the useMobileScreen hook from the @/common/lib/isMobile library. This will allow the component to re-render when the screen size changes.