The diff contains changes to the README.md, package.json, and src/common/components/MetaSEO.tsx files.
In the README.md file, a new line has been added to the list of things to do. This line indicates that SEO adjustment needs to be made.

In the package.json file, a new dependency has been added for the graphql library. This is likely due to the fact that the MetaSEO component now needs to make use of GraphQL in order to fetch data for the SEO tags.

Finally, in the MetaSEO.tsx file, the component has been updated to actually fetch the data from the DatoCMS API and to output the correct SEO tags.