The first diff is for the file src/modules/news/components/Layout.tsx. In this file, the useMobileScreen function was replaced with the useDesktopScreen function. This change was made on line 14.
The second diff is for the file src/modules/scholar/components/ResultHeader.tsx. In this file, the useMobileScreen function was replaced with the useDesktopScreen function. This change was made on line 16.

The third diff is for the file src/modules/scholar/components/ResultSkeleton.tsx. In this file, the useMobileScreen function was replaced with the useDesktopScreen function. This change was made on line 28.