The first diff is for the file HomeMainArticles.tsx. In this file, the author has made changes to the way that articles are filtered and displayed. Previously, the code would only display 5 articles, but now it will display 2 articles if the user is on a desktop screen.
The second diff is for the file getHomeNewsData.ts. In this file, the author has made a change to the way that data is retrieved from the news API. Previously, the code would only retrieve data for the 'svc/topstories/v2/home.json' path, but now it will also retrieve data for the 'search' path.

The third diff is for the file news/index.tsx. In this file, the author has made changes to the layout of the news page. Previously, the news articles would only be displayed in one column, but now they will be displayed in two columns.