The first diff is for the file HomeMainArticles.tsx. In this file, the className for the Link component was changed from text-blue-200 font-semibold text-2xl lg:text-base to text-blue-200 font-semibold text-2xl lg:text-base hover:underline. This change will add a hover effect to the links in the HomeMainArticles component.
The second diff is for the file HomeSideArticles.tsx. In this file, the className for the Link component was changed from text-blue-200 to text-blue-200 hover:underline. This change will add a hover effect to the links in the HomeSideArticles component.

The third diff is for the file getHomeNewsData.ts. In this file, the newsAPI call was changed from q=world to q=world&pageSize=20. This change will make the newsAPI call return 20 results instead of the default 10.

The fourth diff is for the file news/index.tsx. In this file, a new component, NewsHomeWeekendArticles, was added. This component will display articles from the newsAPI call.