The first change is in the ResultSkeleton.tsx file. On line 28, the code changed from !isDesktop to isDesktop. This means that the ScholarResultFilter component will now only be displayed on desktop devices, instead of all devices.
The second change is in the news/index.tsx file. On line 55, a console.log statement was removed. This statement was probably used for debugging purposes.

The third change is in the scholar/index.tsx file. On line 49, the code was changed to use the SERPAPI_URL and SERPAPI_KEY environment variables instead of the NEXT_PUBLIC_BASE_URL variable. This is probably because the SERPAPI_URL provides a more direct link to the search results.