The first thing to note is that the Layout.tsx file has been changed to remove the 'relative' className from the section tag. This could be due to the fact that the author wants to change the position of the header and main sections.
The second change is in the 'py-4 px-2 border-b border-white' className of the header tag. The 'border-white' class has been removed, which could be due to the author wanting to change the color of the border.

The third change is in the 'flex flex-row items-center justify-between' className of the header tag. The 'justify-between' class has been removed, which could be due to the author wanting to change the position of the header elements.

The fourth change is in the 'mr-2' className of the button tag. The 'mr-2' class has been removed, which could be due to the author wanting to change the position of the button.

The fifth change is in the 'text-3xl' className of the FaBuffer and FaUserCircle tags. The 'text-3xl' class has been removed, which could be due to the author wanting to change the size of the text.

The sixth change is in the 'main' tag. The 'min-h-screen' class has been removed, which could be due to the author wanting to change the minimum height of the screen.