This diff contains changes to the Layout.tsx and index.tsx files in the src/modules/home/components directory and the src/pages directory.
In the Layout.tsx file, the header class has been changed from animate-slide-down to just py-4 px-2 border-b text-center font-semibold text-3xl italic. This change will remove the animation from the header element.

In the index.tsx file, a new import, Inter has been added, which will import the Google Inter font. The h1 element's class has been changed to include the Inter font className, as well as the animate-typing class. This will cause the h1 element to use the Inter font and animate the typing effect.

In the tailwind.config.js file, the typing and slide-down animations have been added. The typing animation has been added with a duration of 2 seconds and a linear timing function. The slide-down animation has been added with a duration of 2 seconds and a linear timing function.