The first change we see is in the import of React. Before, React was being imported from the default namespace. Now, it is being imported from the namespace with the useEffect and useState hooks.
The next change is in the Home function. We see that a new state variable, showSearch, is being initialized with a value of false. We also see that a useEffect hook is being used that will set the value of showSearch to true after two seconds.

Finally, we see that a conditional is being added around the SearchBox component. Now, the SearchBox will only be rendered if the value of showSearch is true.