The first change is to the file src/common/components/AdvanceSearchPopup.tsx. This change adds a new class to the element with the class name animate-slide-advance-search. This class will animate the element when it is displayed on the screen.
The second change is to the file src/pages/index.tsx. This change adds a new class to the element with the class name lg:w-1/3. This class will make the element take up 1/3 of the width of the screen on large screens.

The third change is to the file tailwind.config.js. This change adds a new animation to the keyframes object with the name slide-advance-search. This animation will make the element slide down from the top of the screen when it is displayed.