The first thing to note is the import of the Router and classNames modules. Next, we see the addition of a new import, FaGraduationCap, which is a React icon. Finally, we see the addition of a new prop to the IScholarPageLayout interface.
The next section is the start of the header. We see that the header now has a className of py-4 and that it also has a className of flex flex-row items-start border-b pr-2. This means that the header will now be a row of items that are aligned to the start, with a border at the bottom and some padding on the right.

The final section is the main content. We see that the main content now has a className of min-h-screen. This means that the minimum height of the content will be the height of the screen.