The diff contains changes to the AdvanceSearchPopup component and the search page.
In the AdvanceSearchPopup component, the href for the News Data menu item has been updated to include the search value. This will ensure that when the user clicks on the menu item, they are taken to the search page with the correct search value populated.

On the search page, the code has been updated to fetch the news data from the getSearchNewsData function. Previously, the data was being fetched from Redis, but this has been removed.