The first thing to note is that this diff is for multiple files, not just one. The main file being changed here is SearchBox.tsx, but there are also changes to other files in the same directory.
The biggest change in SearchBox.tsx is the addition of a new prop, isLoading. This prop is used to determine whether or not the search icon should be replaced with a spinning loading icon. This is useful for when the search is taking a long time to return results.

There are also changes to the way the search icon is rendered. Previously, the FaSearch icon was used. Now, the FaSpinner icon is used if the isLoading prop is true. This gives the user a better indication that the search is still in progress.