The first change is to the next.config.js file. This change adds a new property to the next.config.js file called publicRuntimeConfig. This property contains two values, BASE_URL and APP_ENV. These values are taken from the environment variables NEXT_PUBLIC_BASE_URL and APP_ENV.
The second change is to the package.json file. This change adds a new dependency called firebase. Firebase is a popular database service that will be used by the app.

The third and final change is to the src/pages/_app.tsx file. This change adds a new import called initFirebaseAnalytic. This import is used to initialize the Firebase Analytics service. This will allow the app to track usage data and analytics.