The first file in this diff is the SearchBox.tsx file. This file has been updated to include a new import, getConfig, as well as a new constant, APP_ENV. These changes allow the SearchBox component to access environment variables, which will be used to determine whether or not to send a firebase event when the search box is used.
The second file in this diff is the NYTSearchResults.tsx file. This file has been updated to include changes to the width and height of the images being rendered. These changes are necessary to ensure that the images are properly sized on the page.

The third file in this diff is the search.tsx file. This file has been updated to include a console.log statement. This statement will be used to debug an issue with the news data being returned from the getSearchNewsData function.