The diff shows the changes made to the src/pages/scholar/index.tsx file.
The first change is to import the MetaSEO component from the @/common/components/MetaSEO directory.

The second change is to import the SerpScholarTable component from the @/modules/scholar/components/SerpScholarTable directory.

The third change is to add a MetaSEO component to the ScholarPageLayout component.

The fourth change is to add a SerpScholarTable component to the ScholarPageLayout component.