The first thing to note is that the diff is for multiple files. The main file being changed is src/pages/scholar/index.tsx. This file is part of the code for the Scholar research data page on the website.
The diff shows that some changes have been made to the way that the page is laid out. In particular, the SearchBox component has been removed and replaced with the ScholarHome component. This has resulted in some changes to the way that the page looks and functions.

Overall, the changes made in this diff improve the usability of the Scholar research data page on the website.