The first change is to the file ScholarSearchSkeleton.tsx. The change here is to the className of a div. The className has been changed from container mx-auto to container mx-auto p-4.
The second change is to the file SerpScholar.tsx. The changes here are to a Button component. The buttonClassName has been changed from flex flex-row justify-between to flex flex-row items-center justify-between p-2 border border-white rounded-sm w-full. In addition, a new state variable has been added called showTable. This variable is used to track whether the table should be shown or not.