The first diff is for the file SerpScholar.tsx. In this file, the author has added a new prop called buttonClassName to the Button component. They have also changed the FaGraduationCap icon to have the class text-2xl and added a new class called mx-2 to the span element that contains the text Google Scholar. Finally, they have added a new div element with the class overflow-y-scroll lg:overflow-hidden and moved the SerpScholarResult component into this div.
The second diff is for the file SerpScholarTable.tsx. In this file, the author has changed the structure of the table header. They have also changed the class name for the table row from flex flex-row items-center to just tr.

The third diff is for the file index.tsx. In this file, the author has removed the assignPageQueryToURL function and the console.log(serpScholar) statement.