The first thing to notice is that the diff is for multiple files. The first file is ScholarSearchSkeleton.tsx, and the second file is SerpScholar.tsx.
In the ScholarSearchSkeleton.tsx file, there are some changes to the way that the component is rendered. The biggest change is that the SerpScholar component is now being rendered without any props.

In the SerpScholar.tsx file, there are some changes to the way that the component is rendered. The biggest change is that the component is now being rendered with a query function. This function will fetch data from the Google Scholar API.

The final file in the diff is the index.tsx file. In this file, there is a change to the way that the ScholarSearchSkeleton component is being rendered. The biggest change is that the component is now being rendered with a query function. This function will fetch data from the Google Scholar API.