The first change is in the Layout.tsx file. This file is responsible for the layout of the Scholar Research page. The change here is to use the useRouter hook from next/router instead of the Router component from next/router. This will allow the search query to be accessed from the router and used in the search input placeholder.
The second change is in the index.tsx file. This file is responsible for rendering the Scholar Research page. The change here is to use the MetaSEO component from @/common/components to render the SEO for the page. This will allow the page to be more easily found by search engines.

The third change is in the getStaticProps function. This function is responsible for fetching the data needed to render the Scholar Research page. The change here is to fetch the data for the SEO from the DatoCMS API. This will allow the page to be more easily found by search engines.