The first thing to note is the addition of the MobileBar component. This component is responsible for displaying a button that, when clicked, will toggle the visibility of the SerpScholarResult component.
Next, we see that the MobileBar component is being used in the SerpScholar component. This component is responsible for displaying the results of a Google Scholar search.

Finally, we see that the SerpScholar component is being given a new prop: showTable. This prop determines whether or not the SerpScholarResult component is visible.