This diff contains changes to the SerpScholar component in the src/modules/scholar/components directory.
The biggest change is the removal of the useQuery and useMutation hooks. These hooks are used to fetch data from an external API and then update the UI accordingly.

The other change is the removal of the SerpScholarResult component. This component was responsible for displaying the data that was fetched from the external API.