The first part of the diff deals with the DesktopBar component. The biggest change here is the addition of a select component which allows the user to filter the results by date. The options for the select component are also being generated dynamically based on the current year.
The second part of the diff deals with the SerpScholarTable component. The biggest change here is that the total number of citations is now being displayed correctly. Previously, it would always say Unknown.

The third and final part of the diff deals with the index component. The biggest change here is that the onFilterChange function is now being passed down to the SerpScholarDesktopBar component.