The first thing to note is that the diff is for multiple files. The main file being changed is the SemanticScholar index.tsx, however there are also changes in the constant.ts and fetchSemanticScholar.ts files.
In the index.tsx file, the biggest change is in the SemanticScholarResult component. This component is responsible for displaying the results of a search query. Previously, it only displayed a loading spinner while the results were being fetched. Now, it also displays a table with the results.

In the constant.ts file, the changes seem to be mostly cosmetic. The labels for the filter options have been changed, and the values have been updated to reflect the current year.

Finally, in the fetchSemanticScholar.ts file, a new parameter has been added to the fetchSemanticScholar function. This parameter is for a filter, which can be used to specify what kind of results you want to see. For example, you can use the filter to only show results from the past year.